Add a polaris.ocean.init_state package for initial-state helpers
Per review feedback, the remaining helpers in
polaris.ocean.vertical.pstar_state are general initial-state utilities,
not p-star specific.  Move them into a new polaris/ocean/init_state/
package (one small module per concern -- thickness, velocity, density --
so the package can grow without bloat) and delete pstar_state.py.
