+++ docs/developers_guide/ocean/tasks/overflow.md
@@ -68,7 +68,7 @@ p-star coordinate to convergence with zero surface pressure, and
 implements `init_tracers()` as the overflow profile (interpreted as
 conservative temperature and absolute salinity) evaluated at the current
 p-star layer midpoints.  After convergence, the shared
-`polaris.ocean.vertical.pstar_state` helpers add layer thickness,
+`polaris.ocean.init_state` helpers add layer thickness,
 quiescent velocity and density, and (for MPAS-Ocean)
 {py:func}`polaris.ocean.eos.convert_tracers_to_mpas_ocean()` converts the
 tracers to potential temperature and practical salinity at a nominal
