+++ polaris/tasks/ocean/overflow/pstar_init.py
@@ -1,13 +1,13 @@
 import xarray as xr
 
 from polaris.ocean.eos import convert_tracers_to_mpas_ocean
-from polaris.ocean.model import OceanIOStep
-from polaris.ocean.vertical.pstar_init import PStarInitStep
-from polaris.ocean.vertical.pstar_state import (
+from polaris.ocean.init_state import (
     add_density_from_specvol,
     add_quiescent_normal_velocity,
     layer_thickness_from_geom_interfaces,
 )
+from polaris.ocean.model import OceanIOStep
+from polaris.ocean.vertical.pstar_init import PStarInitStep
 from polaris.tasks.ocean.overflow.mesh import (
     build_overflow_mesh,
     compute_bottom_depth,
